Move callbacks into function
